src/requirements.txt
Flask==2.0.1
django==4.2.0
gspread==5.0.0
oauth2client==4.1.3
python-dotenv==1.0.0
setuptools>=65.5.1
google-auth==2.28.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.120.0
pillow==10.2.0
python-multipart==0.0.9
whitenoise==6.4.0
openai>=1.76.0
requests>=2.31.0